// src/api/model/Chassis.ts
export interface Chassis {
// ...
    Links?: ChassisLinks;
// ...
}