// src/api/model/EnvironmentMetrics.ts
/**
 * Generated by orval v7.19.0 🍺
 * Do not edit manually.
 * Redfish
 * This contains the definition of a Redfish service.
 * OpenAPI spec version: 2025.4
 */
import type { OdataV4Context } from './OdataV4Context';
import type { OdataV4Etag } from './OdataV4Etag';
import type { OdataV4Id } from './OdataV4Id';
import type { OdataV4Type } from './OdataV4Type';
import type { SensorExcerpt } from './SensorExcerpt';
import type { EnvironmentMetricsActions } from './EnvironmentMetricsActions';
import type { SensorCurrentExcerpt } from './SensorCurrentExcerpt';
import type { ResourceDescription } from './ResourceDescription';
import type { SensorEnergykWhExcerpt } from './SensorEnergykWhExcerpt';
import type { SensorFanArrayExcerpt } from './SensorFanArrayExcerpt';
import type { OdataV4Count } from './OdataV4Count';
import type { ResourceId } from './ResourceId';
import type { ResourceName } from './ResourceName';
import type { ResourceOem } from './ResourceOem';
import type { ControlSingleExcerpt } from './ControlSingleExcerpt';
import type { SensorPowerExcerpt } from './SensorPowerExcerpt';
import type { SensorVoltageExcerpt } from './SensorVoltageExcerpt';

/**
 * The `EnvironmentMetrics` schema represents the environmental metrics of a device.
 * @version 1.5.0
 */
export interface EnvironmentMetrics {
    '@odata.context'?: OdataV4Context;
    '@odata.etag'?: OdataV4Etag;
    '@odata.id': OdataV4Id;
    '@odata.type': OdataV4Type;
    /**
     * Absolute humidity (g/m^3).
     * @nullable
     */
    AbsoluteHumidity?: SensorExcerpt;
    /** The available actions for this resource. */
    Actions?: EnvironmentMetricsActions;
    /**
     * Ambient temperature (Celsius).
     * @nullable
     */
    AmbientTemperatureCelsius?: SensorExcerpt;
    /**
     * The current (A) for this device.
     * @nullable
     */
    CurrentAmps?: SensorCurrentExcerpt;
    /** @nullable */
    readonly Description?: ResourceDescription;
    /**
     * The dew point temperature (C).
     * @nullable
     */
    DewPointCelsius?: SensorExcerpt;
    /**
     * Energy consumption (J).
     * @nullable
     */
    EnergyJoules?: SensorExcerpt;
    /**
     * Energy consumption (kWh).
     * @nullable
     */
    EnergykWh?: SensorEnergykWhExcerpt;
    /** Fan speeds (percent). */
    FanSpeedsPercent?: SensorFanArrayExcerpt[];
    'FanSpeedsPercent@odata.count'?: OdataV4Count;
    /**
     * Humidity (percent).
     * @nullable
     */
    HumidityPercent?: SensorExcerpt;
    readonly Id: ResourceId;
    readonly Name: ResourceName;
    /** The OEM extension property. */
    Oem?: ResourceOem;
    /**
     * Power limit (W).
     * @nullable
     */
    PowerLimitWatts?: ControlSingleExcerpt;
    /**
     * The power load (percent) for this device.
     * @nullable
     */
    PowerLoadPercent?: SensorExcerpt;
    /**
     * Power consumption (W).
     * @nullable
     */
    PowerWatts?: SensorPowerExcerpt;
    /**
     * Temperature (Celsius).
     * @nullable
     */
    TemperatureCelsius?: SensorExcerpt;
    /**
     * The voltage (V) for this device.
     * @nullable
     */
    Voltage?: SensorVoltageExcerpt;
}